send metrics immediately if agents are added or removed
